Rename jbatch.cdi.2.1 to io.openliberty
