--- dev/com.ibm.ws.jbatch.cdi.2.1/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,0 @@
-com.ibm.ws.jbatch.cdi.BatchCDIInjectionExtension
