+++ dev/io.openliberty.jbatch.2.1.cdi/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>com.ibm.ws.jbatch.cdi.2.1</name>
+	<name>io.openliberty.jbatch.cdi.2.1</name>
 	<comment></comment>
 	<projects>
 	</projects>
