+++ dev/io.openliberty.jbatch.2.1.cdi/resources/META-INF/beans.xml
@@ -16,7 +16,7 @@
    version="1.1" 
    bean-discovery-mode="all">
    <scan>
-       <exclude name="com.ibm.ws.jbatch.cdi.**" />
+       <exclude name="io.openliberty.jbatch.cdi.**" />
    </scan>
 </beans>
