+++ dev/io.openliberty.jbatch.2.1.cdi/resources/META-INF/services/jakarta.enterprise.inject.spi.Extension
@@ -1,0 +1,1 @@
+io.openliberty.jbatch.cdi.BatchCDIInjectionExtension
